Extract Tokio code to node/src/tokio.rs
