+++ node/src/lib.rs
@@ -1,7 +1,7 @@
-//! Blockchain Node.
+//! Node.
 
 //
-// Copyright (c) 2019 Stegos AG
+// Copyright (c) 2019-2020 Stegos AG
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -29,21 +29,19 @@ mod error;
 mod mempool;
 pub mod metrics;
 pub mod protos;
+mod tokio;
 mod validation;
+pub use self::tokio::{Node, NodeService};
 pub use crate::api::*;
 pub use crate::config::NodeConfig;
 use crate::error::*;
 use crate::mempool::Mempool;
-use crate::protos::{ChainLoaderMessage, RequestBlocks, ResponseBlocks};
 use crate::validation::*;
 use failure::{bail, format_err, Error};
-use futures::sync::{mpsc, oneshot};
-use futures::{task, Async, AsyncSink, Future, Poll, Sink, Stream};
-use futures_stream_select_all_send::select_all;
+use futures::sync::oneshot;
 use rand::{self, Rng};
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
 use std::collections::{HashMap, HashSet};
-use std::thread;
 use std::time::{Duration, Instant};
 use stegos_blockchain::Timestamp;
 use stegos_blockchain::*;
@@ -55,44 +53,9 @@ use stegos_crypto::hash::Hash;
 use stegos_crypto::scc::Fr;
 use stegos_crypto::vdf::VDF;
 use stegos_crypto::{pbc, scc};
-use stegos_network::{Network, ReplicationEvent};
-use stegos_network::{PeerId, UnicastMessage};
-use stegos_replication::{Replication, ReplicationRow};
 use stegos_serialization::traits::ProtoConvert;
-use stegos_txpool::TransactionPoolService;
 pub use stegos_txpool::MAX_PARTICIPANTS;
-use tokio_timer::{clock, Delay, Interval};
 use Validation::*;
-
-// ----------------------------------------------------------------
-// Public API.
-// ----------------------------------------------------------------
-
-/// Blockchain Node.
-#[derive(Clone, Debug)]
-pub struct Node {
-    outbox: mpsc::UnboundedSender<NodeIncomingEvent>,
-    network: Network,
-}
-
-impl Node {
-    /// Send transaction to node and to the network.
-    pub fn send_transaction(&self, transaction: Transaction) -> oneshot::Receiver<NodeResponse> {
-        let (tx, rx) = oneshot::channel();
-        let request = NodeRequest::BroadcastTransaction { data: transaction };
-        let msg = NodeIncomingEvent::Request { request, tx };
-        self.outbox.unbounded_send(msg).expect("connected");
-        rx
-    }
-
-    /// Execute a Node Request.
-    pub fn request(&self, request: NodeRequest) -> oneshot::Receiver<NodeResponse> {
-        let (tx, rx) = oneshot::channel();
-        let msg = NodeIncomingEvent::Request { request, tx };
-        self.outbox.unbounded_send(msg).expect("connected");
-        rx
-    }
-}
 
 // ----------------------------------------------------------------
 // Internal Implementation.
@@ -101,17 +64,17 @@ impl Node {
 /// Topic used for sending transactions.
 pub const TX_TOPIC: &'static str = "tx";
 /// Topic used for consensus.
-const CONSENSUS_TOPIC: &'static str = "consensus";
+pub(crate) const CONSENSUS_TOPIC: &'static str = "consensus";
 /// Topic for ViewChange message.
-pub const VIEW_CHANGE_TOPIC: &'static str = "view_changes";
+pub(crate) const VIEW_CHANGE_TOPIC: &'static str = "view_changes";
 /// Topic for ViewChange proofs broadcasts.
-pub const VIEW_CHANGE_PROOFS_TOPIC: &'static str = "view_changes_proofs";
+pub(crate) const VIEW_CHANGE_PROOFS_TOPIC: &'static str = "view_changes_proofs";
 /// Topic for ViewChange proofs.
-pub const VIEW_CHANGE_DIRECT: &'static str = "view_changes_direct";
+pub(crate) const VIEW_CHANGE_DIRECT: &'static str = "view_changes_direct";
 /// Topic used for sending sealed blocks.
-const SEALED_BLOCK_TOPIC: &'static str = "block";
+pub(crate) const SEALED_BLOCK_TOPIC: &'static str = "block";
 /// Unicast topic for loading blocks.
-const CHAIN_LOADER_TOPIC: &'static str = "chain-loader";
+pub(crate) const CHAIN_LOADER_TOPIC: &'static str = "chain-loader";
 
 //
 // Logging utils.
@@ -151,10 +114,17 @@ pub enum NodeIncomingEvent {
     Transaction(Vec<u8>),
     Consensus(Vec<u8>),
     Block(Vec<u8>),
+    DecodedBlock(Block),
     ViewChangeMessage(Vec<u8>),
     ViewChangeProof(Vec<u8>),
-    ViewChangeProofMessage(UnicastMessage),
-    ChainLoaderMessage(UnicastMessage),
+    ViewChangeProofMessage {
+        from: pbc::PublicKey,
+        data: Vec<u8>,
+    },
+    ChainLoaderMessage {
+        from: pbc::PublicKey,
+        data: Vec<u8>,
+    },
     CheckSyncTimer,
     MacroBlockProposeTimer,
     MacroBlockViewChangeTimer,
@@ -190,6 +160,14 @@ pub enum NodeOutgoingEvent {
         difficulty: u64,
     },
     MicroBlockViewChangeTimer(Duration),
+    RequestBlocksFrom {
+        from: pbc::PublicKey,
+    },
+    SendBlocksTo {
+        to: pbc::PublicKey,
+        epoch: u64,
+        offset: u32,
+    },
 }
 
 enum MicroBlockTimer {
@@ -226,95 +204,14 @@ enum Validation {
     },
 }
 
-/// Chain subscriber which is fed from the disk.
-struct ChainReader {
-    /// Current epoch.
-    epoch: u64,
-    /// Current offset.
-    offset: u32,
-    /// Channel.
-    tx: mpsc::Sender<ChainNotification>,
-}
-
-impl ChainReader {
-    fn poll(&mut self, chain: &Blockchain) -> Poll<(), Error> {
-        // Check if subscriber has already been synchronized.
-        if self.epoch == chain.epoch() && self.offset == chain.offset() {
-            return Ok(Async::Ready(()));
-        }
-
-        // Feed blocks from the disk.
-        for block in chain.blocks_starting(self.epoch, self.offset) {
-            let (msg, next_epoch, next_offset) = match block {
-                Block::MacroBlock(block) => {
-                    assert_eq!(block.header.epoch, self.epoch);
-                    let epoch_info = chain.epoch_info(block.header.epoch)?.unwrap().clone();
-                    let next_epoch = block.header.epoch + 1;
-                    let msg = ExtendedMacroBlock { block, epoch_info };
-                    let msg = ChainNotification::MacroBlockCommitted(msg);
-                    (msg, next_epoch, 0)
-                }
-                Block::MicroBlock(block) => {
-                    assert_eq!(block.header.epoch, self.epoch);
-                    assert_eq!(block.header.offset, self.offset);
-                    let (next_epoch, next_offset) =
-                        if block.header.offset + 1 < chain.cfg().micro_blocks_in_epoch {
-                            (block.header.epoch, block.header.offset + 1)
-                        } else {
-                            (block.header.epoch + 1, 0)
-                        };
-                    let msg = ChainNotification::MicroBlockPrepared(block);
-                    (msg, next_epoch, next_offset)
-                }
-            };
-
-            match self.tx.start_send(msg)? {
-                AsyncSink::Ready => {
-                    self.epoch = next_epoch;
-                    self.offset = next_offset;
-                }
-                AsyncSink::NotReady(_msg) => {
-                    break;
-                }
-            }
-        }
-
-        self.tx.poll_complete()?;
-        Ok(Async::NotReady)
-    }
-}
-
-/// Notify all subscribers about new event.
-fn notify_subscribers<T: Clone>(subscribers: &mut Vec<mpsc::Sender<T>>, msg: T) {
-    let mut i = 0;
-    while i < subscribers.len() {
-        let tx = &mut subscribers[i];
-        match tx.start_send(msg.clone()) {
-            Ok(AsyncSink::Ready) => {}
-            Ok(AsyncSink::NotReady(_msg)) => {
-                log::warn!("Subscriber is slow, discarding messages");
-            }
-            Err(_e /* SendError<ChainNotification> */) => {
-                subscribers.swap_remove(i);
-                continue;
-            }
-        }
-        if let Err(_e) = tx.poll_complete() {
-            subscribers.swap_remove(i);
-            continue;
-        }
-        i += 1;
-    }
-}
-
-pub struct NodeService {
+pub struct NodeState {
     /// Config.
-    cfg: NodeConfig,
+    pub(crate) cfg: NodeConfig,
     chain_name: String,
     /// Blockchain.
-    chain: Blockchain,
+    pub(crate) chain: Blockchain,
     /// Network secret key.
-    network_pkey: pbc::PublicKey,
+    pub(crate) network_pkey: pbc::PublicKey,
     /// Network secret key.
     network_skey: pbc::SecretKey,
 
@@ -336,55 +233,21 @@ pub struct NodeService {
     /// Automatic re-staking status.
     is_restaking_enabled: bool,
 
-    outgoing: Vec<NodeOutgoingEvent>,
-
-    //----------------------------------------------------------------------------------------------
-    // Tokio-part
-    //----------------------------------------------------------------------------------------------
-    /// Timer to check sync status
-    check_sync: Interval,
-
-    /// Aggregated stream of events.
-    events: Box<dyn Stream<Item = NodeIncomingEvent, Error = ()> + Send>,
-
-    /// Subscribers for status events.
-    status_subscribers: Vec<mpsc::Sender<StatusNotification>>,
-    /// Subscribers for chain events.
-    chain_subscribers: Vec<mpsc::Sender<ChainNotification>>,
-    /// Subscribers for chain events which are fed from the disk.
-    /// Automatically promoted to chain_subscribers after synchronization.
-    chain_readers: Vec<ChainReader>,
-    /// Network interface.
-    network: Network,
-
-    macro_block_propose_timer: Option<Delay>,
-    macro_block_view_change_timer: Option<Delay>,
-    micro_block_propose_timer: Option<oneshot::Receiver<Vec<u8>>>,
-    micro_block_view_change_timer: Option<Delay>,
-
-    /// Txpool
-    txpool_service: Option<TransactionPoolService>,
-
-    /// Replication
-    replication: Replication,
+    pub(crate) outgoing: Vec<NodeOutgoingEvent>,
 }
 
-impl NodeService {
+impl NodeState {
     /// Constructor.
     pub fn new(
         cfg: NodeConfig,
         chain: Blockchain,
         network_skey: pbc::SecretKey,
         network_pkey: pbc::PublicKey,
-        network: Network,
         chain_name: String,
-        peer_id: PeerId,
-        replication_rx: mpsc::UnboundedReceiver<ReplicationEvent>,
-    ) -> Result<(Self, Node), Error> {
-        let (outbox, inbox) = mpsc::unbounded();
+    ) -> Result<Self, Error> {
         let mempool = Mempool::new();
 
-        let last_block_clock = clock::now();
+        let last_block_clock = Instant::now();
         let validation = if chain.is_epoch_full() {
             MacroBlockAuditor
         } else {
@@ -395,67 +258,7 @@ impl NodeService {
         let restaking_offset = 0; // will be updated on init().
         let is_restaking_enabled = true;
 
-        let status_subscribers = Vec::new();
-
-        let mut streams =
-            Vec::<Box<dyn Stream<Item = NodeIncomingEvent, Error = ()> + Send>>::new();
-
-        // Control messages
-        streams.push(Box::new(inbox));
-
-        // Transaction Requests
-        let transaction_rx = network
-            .subscribe(&TX_TOPIC)?
-            .map(|m| NodeIncomingEvent::Transaction(m));
-        streams.push(Box::new(transaction_rx));
-
-        // Consensus Requests
-        let consensus_rx = network
-            .subscribe(&CONSENSUS_TOPIC)?
-            .map(|m| NodeIncomingEvent::Consensus(m));
-        streams.push(Box::new(consensus_rx));
-
-        let view_change_rx = network
-            .subscribe(&VIEW_CHANGE_TOPIC)?
-            .map(|m| NodeIncomingEvent::ViewChangeMessage(m));
-        streams.push(Box::new(view_change_rx));
-
-        let view_change_proofs_rx = network
-            .subscribe(&VIEW_CHANGE_PROOFS_TOPIC)?
-            .map(|m| NodeIncomingEvent::ViewChangeProof(m));
-        streams.push(Box::new(view_change_proofs_rx));
-
-        let view_change_unicast_rx = network
-            .subscribe_unicast(&VIEW_CHANGE_DIRECT)?
-            .map(|m| NodeIncomingEvent::ViewChangeProofMessage(m));
-        streams.push(Box::new(view_change_unicast_rx));
-
-        // Sealed blocks broadcast topic.
-        let block_rx = network
-            .subscribe(&SEALED_BLOCK_TOPIC)?
-            .map(|m| NodeIncomingEvent::Block(m));
-        streams.push(Box::new(block_rx));
-
-        // Chain loader messages.
-        let requests_rx = network
-            .subscribe_unicast(CHAIN_LOADER_TOPIC)?
-            .map(NodeIncomingEvent::ChainLoaderMessage);
-        streams.push(Box::new(requests_rx));
-
-        let events = select_all(streams);
-
-        let check_sync = Interval::new_interval(cfg.sync_change_timeout);
-        let chain_readers = Vec::new();
-        let chain_subscribers = Vec::new();
-        let node = Node {
-            outbox,
-            network: network.clone(),
-        };
-        let txpool_service = None;
-        let light = false;
-        let replication = Replication::new(peer_id, network.clone(), light, replication_rx);
-
-        let service = NodeService {
+        let state = NodeState {
             cfg,
             chain_name,
             chain,
@@ -468,23 +271,10 @@ impl NodeService {
             restaking_offset,
             is_restaking_enabled,
             outgoing: Vec::new(),
-
-            chain_readers,
-            chain_subscribers,
-            network: network.clone(),
-            macro_block_propose_timer: None,
-            macro_block_view_change_timer: None,
-            micro_block_propose_timer: None,
-            micro_block_view_change_timer: None,
-            check_sync,
-            events,
-            txpool_service,
-            replication,
-            status_subscribers,
         };
-        service.update_stake_balance();
+        state.update_stake_balance();
 
-        Ok((service, node))
+        Ok(state)
     }
 
     /// Invoked when network is ready.
@@ -827,7 +617,11 @@ impl NodeService {
                 "Found a fork with lower view_change, sending blocks: pkey={}",
                 remote.header.pkey
             );
-            self.send_blocks(remote.header.pkey, epoch, offset)?;
+            self.outgoing.push(NodeOutgoingEvent::SendBlocksTo {
+                to: remote.header.pkey,
+                epoch,
+                offset,
+            });
             return Err(ForkError::Canceled);
         }
 
@@ -914,7 +708,8 @@ impl NodeService {
                   local_view_change,
                   remote_view_change);
             // Request history from that node.
-            self.request_history_from(pkey, "fork resolution")?;
+            self.outgoing
+                .push(NodeOutgoingEvent::RequestBlocksFrom { from: pkey });
             return Err(ForkError::Canceled);
         }
 
@@ -1068,7 +863,8 @@ impl NodeService {
                 Ok(BlockchainError::BlockError(BlockError::InvalidMicroBlockPreviousHash(..))) => {
                     // A potential fork - request history from that node.
                     let from = self.chain.select_leader(view_change);
-                    self.request_history_from(from, "invalid previous hash")?;
+                    self.outgoing
+                        .push(NodeOutgoingEvent::RequestBlocksFrom { from });
                 }
                 Ok(BlockchainError::BlockError(BlockError::InvalidViewChange(..))) => {
                     assert!(self.chain.view_change() > 0);
@@ -1188,7 +984,7 @@ impl NodeService {
             "Validating a micro block: epoch={}, offset={}, block={}, inputs_len={}, outputs_len={}, txs_len={}",
             epoch, offset, &hash, inputs_len, outputs_len, txs_len
         );
-        let start_clock = clock::now();
+        let start_clock = Instant::now();
         let r = {
             let mut outputs: Vec<&Output> = Vec::new();
             for tx in &block.transactions {
@@ -1216,7 +1012,7 @@ impl NodeService {
                 Err(e) => Err(e),
             }
         };
-        let duration = clock::now().duration_since(start_clock);
+        let duration = Instant::now().duration_since(start_clock);
         let duration = (duration.as_secs() as f64) + (duration.subsec_nanos() as f64) * 1e-9;
         metrics::MICRO_BLOCK_VALIDATE_TIME.set(duration);
         metrics::MICRO_BLOCK_VALIDATE_TIME_HG.observe(duration);
@@ -1262,7 +1058,7 @@ impl NodeService {
         metrics::MEMPOOL_INPUTS.set(self.mempool.inputs_len() as i64);
         metrics::MEMPOOL_OUTPUTS.set(self.mempool.inputs_len() as i64);
         let last_block_clock = self.last_block_clock;
-        self.last_block_clock = clock::now();
+        self.last_block_clock = Instant::now();
         let local_timestamp: f64 = Timestamp::now().into();
         let remote_timestamp: f64 = block_timestamp.into();
         let lag = local_timestamp - remote_timestamp; // can be negative.
@@ -1333,33 +1129,6 @@ impl NodeService {
             .push(NodeOutgoingEvent::StatusNotification(msg.into()));
     }
 
-    /// Handler subscription to status.
-    fn handle_subscription_to_status(
-        &mut self,
-    ) -> Result<mpsc::Receiver<StatusNotification>, Error> {
-        let (tx, rx) = mpsc::channel(1);
-        self.status_subscribers.push(tx);
-        Ok(rx)
-    }
-
-    /// Handle subscription to chain.
-    fn handle_subscription_to_chain(
-        &mut self,
-        epoch: u64,
-        offset: u32,
-    ) -> Result<mpsc::Receiver<ChainNotification>, Error> {
-        if epoch > self.chain.epoch() {
-            return Err(format_err!("Invalid epoch requested: epoch={}", epoch));
-        }
-        // Set buffer size to fit entire epoch plus some extra blocks.
-        let buffer = self.chain.cfg().micro_blocks_in_epoch as usize + 10;
-        let (tx, rx) = mpsc::channel(buffer);
-        let subscriber = ChainReader { tx, epoch, offset };
-        self.chain_readers.push(subscriber);
-        task::current().notify();
-        Ok(rx)
-    }
-
     /// Handler for NodeRequest::BroadcastTransaction
     fn handle_add_tx(&mut self, tx: Transaction) -> Result<TransactionStatus, Error> {
         match self.send_transaction(tx.clone()) {
@@ -1393,7 +1162,7 @@ impl NodeService {
     ///
     fn pop_micro_block(&mut self) -> Result<(), Error> {
         let (pruned_outputs, recovered_inputs, txs, block) = self.chain.pop_micro_block()?;
-        self.last_block_clock = clock::now();
+        self.last_block_clock = Instant::now();
         self.mempool.pop_micro_block(txs);
 
         // Update validation status.
@@ -1683,14 +1452,14 @@ impl NodeService {
                 epoch,
                 &block_hash
             );
-            let start_clock = clock::now();
+            let start_clock = Instant::now();
             let r = self.chain.validate_proposed_macro_block(
                 view_change,
                 block_hash,
                 &block_proposal.header,
                 &block_proposal.transactions,
             );
-            let duration = clock::now().duration_since(start_clock);
+            let duration = Instant::now().duration_since(start_clock);
             let duration = (duration.as_secs() as f64) + (duration.subsec_nanos() as f64) * 1e-9;
             metrics::MACRO_BLOCK_VALIDATE_TIME.set(duration);
             metrics::MACRO_BLOCK_VALIDATE_TIME_HG.observe(duration);
@@ -1771,7 +1540,7 @@ impl NodeService {
             self.chain.epoch(),
             consensus.round()
         );
-        let start_clock = clock::now();
+        let start_clock = Instant::now();
 
         // Propose a new block.
         let recipient_pkey = self
@@ -1794,7 +1563,7 @@ impl NodeService {
             transactions,
         };
 
-        let duration = clock::now().duration_since(start_clock);
+        let duration = Instant::now().duration_since(start_clock);
         let duration = (duration.as_secs() as f64) + (duration.subsec_nanos() as f64) * 1e-9;
         metrics::MACRO_BLOCK_CREATE_TIME.set(duration);
         metrics::MACRO_BLOCK_CREATE_TIME_HG.observe(duration);
@@ -1815,7 +1584,9 @@ impl NodeService {
 
     /// Checks if it's time to perform a view change on a micro block.
     fn handle_macro_block_viewchange_timer(&mut self) -> Result<(), Error> {
-        assert!(clock::now().duration_since(self.last_block_clock) >= self.cfg.macro_block_timeout);
+        assert!(
+            Instant::now().duration_since(self.last_block_clock) >= self.cfg.macro_block_timeout
+        );
 
         // Check that a block has been committed but haven't send by the leader.
         let (consensus, block_timer, autocommit) = match &mut self.validation {
@@ -1933,7 +1704,9 @@ impl NodeService {
                     "Received an invalid view_change message: view_change={}, validator={}, error={}",
                     msg.chain.view_change, validator_pkey, e
                 );
-                self.request_history_from(validator_pkey, "invalid view change")?;
+                self.outgoing.push(NodeOutgoingEvent::RequestBlocksFrom {
+                    from: validator_pkey,
+                });
             }
             Err(e) => return Err(e.into()),
         }
@@ -1943,7 +1716,7 @@ impl NodeService {
 
     /// Checks if it's time to perform a view change on a micro block.
     fn handle_micro_block_viewchange_timer(&mut self) -> Result<(), Error> {
-        let elapsed = clock::now().duration_since(self.last_block_clock);
+        let elapsed = Instant::now().duration_since(self.last_block_clock);
         assert!(elapsed >= self.cfg.micro_block_timeout);
         let leader = self.chain.leader();
         swarn!(
@@ -2042,7 +1815,7 @@ impl NodeService {
             view_change,
             previous
         );
-        let start_clock = clock::now();
+        let start_clock = Instant::now();
 
         for (cheater, proof) in &self.cheating_proofs {
             // the cheater was already punished, so we keep proofs for rollback case,
@@ -2084,7 +1857,7 @@ impl NodeService {
         // Sign block.
         block.sign(&self.network_skey, &self.network_pkey);
 
-        let duration = clock::now().duration_since(start_clock);
+        let duration = Instant::now().duration_since(start_clock);
         let duration = (duration.as_secs() as f64) + (duration.subsec_nanos() as f64) * 1e-9;
         metrics::MICRO_BLOCK_CREATE_TIME.set(duration);
         metrics::MICRO_BLOCK_CREATE_TIME_HG.observe(duration);
@@ -2200,149 +1973,6 @@ impl NodeService {
 
         let tx = PaymentTransaction::new(&secret_key, &inputs, &outputs, &outputs_gamma, fee)?;
         Ok(tx.into())
-    }
-
-    /////////////////////////////////////////////////////////////////////////////////////////////////
-    // Loader
-    /////////////////////////////////////////////////////////////////////////////////////////////////
-
-    pub fn request_history_from(
-        &mut self,
-        from: pbc::PublicKey,
-        reason: &str,
-    ) -> Result<(), Error> {
-        let epoch = self.chain.epoch();
-        sinfo!(
-            self,
-            "Downloading blocks: from={}, epoch={}, reason='{}'",
-            &from,
-            epoch,
-            reason
-        );
-        let msg = ChainLoaderMessage::Request(RequestBlocks::new(epoch));
-        self.network
-            .send(from, CHAIN_LOADER_TOPIC, msg.into_buffer()?)
-    }
-
-    fn handle_request_blocks(
-        &mut self,
-        pkey: pbc::PublicKey,
-        request: RequestBlocks,
-    ) -> Result<(), Error> {
-        if request.epoch > self.chain.epoch() {
-            swarn!(
-                self,
-                "Received a loader request with epoch >= our_epoch: remote_epoch={}, our_epoch={}",
-                request.epoch,
-                self.chain.epoch()
-            );
-            return Ok(());
-        }
-
-        self.send_blocks(pkey, request.epoch, 0)
-    }
-
-    pub fn send_blocks(
-        &mut self,
-        pkey: pbc::PublicKey,
-        epoch: u64,
-        offset: u32,
-    ) -> Result<(), Error> {
-        let mut blocks: Vec<Block> = Vec::new();
-        for block in self.chain.blocks_starting(epoch, offset) {
-            blocks.push(block);
-            // Feed the whole epoch.
-            match blocks.last().unwrap() {
-                Block::MacroBlock(_) if blocks.len() > 0 => {
-                    break;
-                }
-                Block::MacroBlock(_) => {}
-                Block::MicroBlock(_) => {}
-            }
-        }
-        sinfo!(
-            self,
-            "Feeding blocks: to={}, num_blocks={}",
-            pkey,
-            blocks.len()
-        );
-        let msg = ChainLoaderMessage::Response(ResponseBlocks::new(blocks));
-        self.network
-            .send(pkey, CHAIN_LOADER_TOPIC, msg.into_buffer()?)?;
-        Ok(())
-    }
-
-    fn handle_response_blocks(
-        &mut self,
-        pkey: pbc::PublicKey,
-        response: ResponseBlocks,
-    ) -> Result<(), Error> {
-        let first_epoch = match response.blocks.first() {
-            Some(Block::MacroBlock(block)) => block.header.epoch,
-            Some(Block::MicroBlock(block)) => block.header.epoch,
-            None => {
-                // Empty response
-                sinfo!(
-                    self,
-                    "Received blocks: from={}, num_blocks={}",
-                    pkey,
-                    response.blocks.len()
-                );
-                return Ok(());
-            }
-        };
-        let last_epoch = match response.blocks.last() {
-            Some(Block::MacroBlock(block)) => block.header.epoch,
-            Some(Block::MicroBlock(block)) => block.header.epoch,
-            None => unreachable!("Checked above"),
-        };
-        if first_epoch > self.chain.epoch() {
-            swarn!(
-                self,
-                "Received blocks from the future: from={}, our_epoch={}, first_epoch={}",
-                pkey,
-                self.chain.epoch(),
-                first_epoch
-            );
-            return Ok(());
-        } else if last_epoch < self.chain.epoch() {
-            swarn!(
-                self,
-                "Received blocks from the past: from={}, last_epoch={}, our_epoch={}",
-                pkey,
-                last_epoch,
-                self.chain.epoch()
-            );
-            return Ok(());
-        }
-
-        sinfo!(
-            self,
-            "Received blocks: from={}, first_epoch={}, our_epoch={}, last_epoch={}, num_blocks={}",
-            pkey,
-            first_epoch,
-            self.chain.epoch(),
-            last_epoch,
-            response.blocks.len()
-        );
-
-        for block in response.blocks {
-            // Fail on the first error.
-            self.handle_block(block)?;
-        }
-
-        Ok(())
-    }
-
-    pub fn handle_chain_loader_message(
-        &mut self,
-        pkey: pbc::PublicKey,
-        msg: ChainLoaderMessage,
-    ) -> Result<(), Error> {
-        match msg {
-            ChainLoaderMessage::Request(r) => self.handle_request_blocks(pkey, r),
-            ChainLoaderMessage::Response(r) => self.handle_response_blocks(pkey, r),
-        }
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////
@@ -2557,17 +2187,14 @@ impl NodeService {
                 .and_then(|proof| {
                     self.handle_view_change_direct(proof.view_change_proof, proof.pkey)
                 }),
-            NodeIncomingEvent::ViewChangeProofMessage(msg) => {
-                SealedViewChangeProof::from_buffer(&msg.data)
-                    .and_then(|proof| self.handle_view_change_direct(proof, msg.from))
+            NodeIncomingEvent::ViewChangeProofMessage { from, data } => {
+                SealedViewChangeProof::from_buffer(&data)
+                    .and_then(|proof| self.handle_view_change_direct(proof, from))
             }
             NodeIncomingEvent::Block(msg) => {
                 Block::from_buffer(&msg).and_then(|msg| self.handle_block(msg))
             }
-            NodeIncomingEvent::ChainLoaderMessage(msg) => {
-                ChainLoaderMessage::from_buffer(&msg.data)
-                    .and_then(|data| self.handle_chain_loader_message(msg.from, data))
-            }
+            NodeIncomingEvent::DecodedBlock(msg) => self.handle_block(msg),
             NodeIncomingEvent::CheckSyncTimer => {
                 if !self.chain.is_synchronized() {
                     self.on_status_changed();
@@ -2590,271 +2217,12 @@ impl NodeService {
                 //
                 self.handle_micro_block_viewchange_timer()
             }
+            NodeIncomingEvent::ChainLoaderMessage { from: _, data: _ } => {
+                unreachable!("Must be handled by NodeService");
+            }
         };
         if let Err(e) = result {
             serror!(self, "Error: {}", e);
         }
-    }
-}
-
-// Event loop.
-impl Future for NodeService {
-    type Item = ();
-    type Error = ();
-
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        if let Some(timer) = &mut self.macro_block_propose_timer {
-            match timer.poll().unwrap() {
-                Async::Ready(()) => {
-                    std::mem::replace(&mut self.macro_block_propose_timer, None);
-                    let event = NodeIncomingEvent::MacroBlockProposeTimer;
-                    self.handle_event(event);
-                }
-                Async::NotReady => {}
-            }
-        }
-        if let Some(timer) = &mut self.macro_block_view_change_timer {
-            match timer.poll().unwrap() {
-                Async::Ready(()) => {
-                    std::mem::replace(&mut self.macro_block_view_change_timer, None);
-                    let event = NodeIncomingEvent::MacroBlockViewChangeTimer;
-                    self.handle_event(event);
-                }
-                Async::NotReady => {}
-            }
-        }
-        if let Some(solver) = &mut self.micro_block_propose_timer {
-            match solver.poll().unwrap() {
-                Async::Ready(solution) => {
-                    std::mem::replace(&mut self.micro_block_propose_timer, None);
-                    let event = NodeIncomingEvent::MicroBlockProposeTimer(solution);
-                    self.handle_event(event);
-                }
-                Async::NotReady => {}
-            }
-        }
-        if let Some(timer) = &mut self.micro_block_view_change_timer {
-            match timer.poll().unwrap() {
-                Async::Ready(()) => {
-                    std::mem::replace(&mut self.micro_block_view_change_timer, None);
-                    let event = NodeIncomingEvent::MicroBlockViewChangeTimer;
-                    self.handle_event(event);
-                }
-                Async::NotReady => {}
-            }
-        }
-
-        loop {
-            match self.check_sync.poll() {
-                Ok(Async::Ready(Some(_))) => {
-                    let event = NodeIncomingEvent::CheckSyncTimer;
-                    self.handle_event(event);
-                    break;
-                }
-                Ok(Async::Ready(None)) => {
-                    serror!(self, "Error during process sync status");
-                    return Ok(Async::Ready(()));
-                }
-                Err(e) => {
-                    serror!(self, "Error: {}", e);
-                    return Err(());
-                }
-                Ok(Async::NotReady) => {
-                    break;
-                }
-            }
-        }
-
-        // Poll chain readers.
-        let mut i = 0;
-        while i < self.chain_readers.len() {
-            match self.chain_readers[i].poll(&self.chain) {
-                Ok(Async::Ready(())) => {
-                    // Synchronized with node, convert into a subscription.
-                    let subscriber = self.chain_readers.swap_remove(i);
-                    self.chain_subscribers.push(subscriber.tx);
-                }
-                Ok(Async::NotReady) => {
-                    i += 1;
-                }
-                Err(_e) => {
-                    self.chain_readers.swap_remove(i);
-                }
-            }
-        }
-
-        if let Some(ref mut txpool_service) = &mut self.txpool_service {
-            match txpool_service.poll().unwrap() {
-                Async::Ready(()) => return Ok(Async::Ready(())), // Shutdown.
-                Async::NotReady => {}
-            };
-        }
-        // Poll internal events.
-        loop {
-            match self.events.poll().expect("all errors are already handled") {
-                Async::Ready(Some(event)) => {
-                    match event {
-                        NodeIncomingEvent::Request { request, tx } => {
-                            match request {
-                                NodeRequest::ChangeUpstream {} => {
-                                    self.replication.change_upstream();
-                                    let response = NodeResponse::UpstreamChanged;
-                                    tx.send(response).ok(); // ignore errors.
-                                    continue;
-                                }
-                                NodeRequest::ReplicationInfo {} => {
-                                    let response =
-                                        NodeResponse::ReplicationInfo(self.replication.info());
-                                    tx.send(response).ok(); // ignore errors.
-                                    continue;
-                                }
-                                NodeRequest::SubscribeChain { epoch, offset } => {
-                                    let response =
-                                        match self.handle_subscription_to_chain(epoch, offset) {
-                                            Ok(rx) => NodeResponse::SubscribedChain {
-                                                current_epoch: self.chain.epoch(),
-                                                current_offset: self.chain.offset(),
-                                                rx: Some(rx),
-                                            },
-                                            Err(e) => NodeResponse::Error {
-                                                error: format!("{}", e),
-                                            },
-                                        };
-                                    tx.send(response).ok(); // ignore errors.
-                                    continue;
-                                }
-                                NodeRequest::SubscribeStatus {} => {
-                                    let response = match self.handle_subscription_to_status() {
-                                        Ok(rx) => {
-                                            let status = self.chain.status();
-                                            NodeResponse::SubscribedStatus {
-                                                status,
-                                                rx: Some(rx),
-                                            }
-                                        }
-                                        Err(e) => NodeResponse::Error {
-                                            error: format!("{}", e),
-                                        },
-                                    };
-                                    tx.send(response).ok(); // ignore errors.
-                                    continue;
-                                }
-                                request => {
-                                    let event = NodeIncomingEvent::Request { request, tx };
-                                    self.handle_event(event)
-                                }
-                            }
-                        }
-                        event => self.handle_event(event),
-                    }
-                }
-                Async::Ready(None) => return Ok(Async::Ready(())), // Shutdown.
-                Async::NotReady => break,
-            }
-        }
-        // Replication
-        loop {
-            let micro_blocks_in_epoch = self.chain.cfg().micro_blocks_in_epoch;
-            let block_reader: &dyn BlockReader = &self.chain;
-            match self.replication.poll(
-                self.chain.epoch(),
-                self.chain.offset(),
-                micro_blocks_in_epoch,
-                block_reader,
-            ) {
-                Async::Ready(Some(ReplicationRow::LightBlock(_block))) => {
-                    panic!("Received the light block from the replication");
-                }
-                Async::Ready(Some(ReplicationRow::Block(block))) => {
-                    if let Err(e) = self.handle_block(block) {
-                        serror!(self, "Invalid block received from replication: {}", e);
-                    }
-                }
-                Async::Ready(None) => return Ok(Async::Ready(())), // Shutdown.
-                Async::NotReady => break,
-            }
-        }
-
-        //
-        // Flush events.
-        //
-        for event in std::mem::replace(&mut self.outgoing, Vec::new()) {
-            let result = match event {
-                NodeOutgoingEvent::FacilitatorChanged { facilitator } => {
-                    if facilitator == self.network_pkey {
-                        sinfo!(self, "I am facilitator");
-                        let txpool_service = TransactionPoolService::new(self.network.clone());
-                        self.txpool_service = Some(txpool_service);
-                    } else {
-                        sinfo!(self, "Facilitator is {}", facilitator);
-                        self.txpool_service = None;
-                    }
-                    Ok(())
-                }
-                NodeOutgoingEvent::Publish { topic, data } => {
-                    //
-                    self.network.publish(&topic, data)
-                }
-                NodeOutgoingEvent::Send { dest, topic, data } => {
-                    //
-                    self.network.send(dest, &topic, data)
-                }
-                NodeOutgoingEvent::MacroBlockProposeTimer(duration) => {
-                    let deadline = clock::now() + duration;
-                    self.macro_block_propose_timer = Some(Delay::new(deadline));
-                    task::current().notify();
-                    Ok(())
-                }
-                NodeOutgoingEvent::MacroBlockViewChangeTimer(duration) => {
-                    let deadline = clock::now() + duration;
-                    self.macro_block_view_change_timer = Some(Delay::new(deadline));
-                    task::current().notify();
-                    Ok(())
-                }
-                NodeOutgoingEvent::MicroBlockProposeTimer {
-                    random,
-                    vdf,
-                    difficulty,
-                } => {
-                    let (tx, rx) = oneshot::channel::<Vec<u8>>();
-                    let challenge = random.to_bytes();
-                    let solver = move || {
-                        let solution = vdf.solve(&challenge, difficulty);
-                        tx.send(solution).ok(); // ignore errors.
-                    };
-                    // Spawn a background thread to solve VDF puzzle.
-                    thread::spawn(solver);
-                    self.micro_block_propose_timer = Some(rx);
-                    task::current().notify();
-                    Ok(())
-                }
-                NodeOutgoingEvent::MicroBlockViewChangeTimer(duration) => {
-                    let deadline = clock::now() + duration;
-                    self.micro_block_view_change_timer = Some(Delay::new(deadline));
-                    task::current().notify();
-                    Ok(())
-                }
-                NodeOutgoingEvent::ReplicationBlock { block, light_block } => {
-                    self.replication.on_block(
-                        block,
-                        light_block,
-                        self.chain.cfg().micro_blocks_in_epoch,
-                    );
-                    Ok(())
-                }
-                NodeOutgoingEvent::StatusNotification(notification) => {
-                    notify_subscribers(&mut self.status_subscribers, notification);
-                    Ok(())
-                }
-                NodeOutgoingEvent::ChainNotification(notification) => {
-                    notify_subscribers(&mut self.chain_subscribers, notification);
-                    Ok(())
-                }
-            };
-            if let Err(e) = result {
-                serror!(self, "Error: {}", e);
-            }
-        }
-        Ok(Async::NotReady)
     }
 }
